Structured_output_parser_enforced_schema.py
from langchain_huggingface import ChatHuggingFace, HuggingFaceEndpoint
from langchain_core.prompts import PromptTemplate
from langchain.output_parsers import StructuredOutputParser, ResponseSchema
from dotenv import load_dotenv

## This is used for defining schema of JSON by yourself which is not supported by json parser
## This does not provides data validation

load_dotenv()

llm = HuggingFaceEndpoint(repo_id="sarvamai/sarvam-m", task="text-generation")

model = ChatHuggingFace(llm=llm)

schema = [
    ResponseSchema(name="fact_1", description="Fact 1 about the topic"),
    ResponseSchema(name="fact_2", description="Fact 2 about the topic"),
    ResponseSchema(name="fact_3", description="Fact 3 about the topic"),
]

parser = StructuredOutputParser.from_response_schemas(schema)

template = PromptTemplate(
    template="Give 3 fact about {topic} \n {format_instruction}",
    input_variables=["topic"],
    partial_variables={"format_instruction": parser.get_format_instructions()},
)

prompt = template.invoke({"topic": "RAG"})

chain = template | model | parser

chain.invoke({"topic": "RAG"})
